add new data module with balance_rank_transform and quantile_bin functions for building crypto data, add unit tests for new functions, refactor existing functions into more appropriately named modules and avoid circular imports
