+++ numerai_tools/typing.py
@@ -1,0 +1,6 @@
+from typing import Union, TypeVar
+
+import pandas as pd
+
+S1 = TypeVar("S1", bound=Union[pd.DataFrame, pd.Series])
+S2 = TypeVar("S2", bound=Union[pd.DataFrame, pd.Series])
