+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "numerai-tools"
-version = "0.5.1"
+version = "0.6.0.dev0"
 description = "A collection of open-source tools to help interact with Numerai, model data, and automate submissions."
 authors = [
     {name = "Numerai Engineering",email = "engineering@numer.ai"}
